refactor: reorganize connection types and improve media handling in downloadMediaFromMessages
