+++ src/baileys/types.ts
@@ -1,0 +1,17 @@
+import type { BaileysEventMap } from "@whiskeysockets/baileys";
+
+export interface BaileysConnectionOptions {
+  clientName?: string;
+  phoneNumber: string;
+  webhookUrl: string;
+  webhookVerifyToken: string;
+  isReconnect?: boolean;
+  includeMedia?: boolean;
+  onConnectionClose?: () => void;
+}
+
+export interface BaileysConnectionWebhookPayload {
+  event: keyof BaileysEventMap;
+  data: BaileysEventMap[keyof BaileysEventMap] | { error: string };
+  extra?: unknown;
+}
